move otf compilation code to otf.compile
