+++ src/functional/otf/compile/build_systems/compiledb.py
@@ -4,15 +4,15 @@ import pathlib
 import subprocess
 from typing import Optional
 
-import functional.otf.stages
 from functional.otf import languages, source, stages, step_types
-from functional.program_processors.builders import build_data, cache, otf_compiler
-from functional.program_processors.builders.cpp import cmake, cmake_lists
+from functional.otf.compile import build_data, compiler
+from functional.otf.compile.build_systems import cmake, cmake_lists
+from functional.program_processors.builders import cache
 
 
 @dataclasses.dataclass()
 class Compiledb(
-    functional.otf.stages.BuildSystemProject[
+    stages.BuildSystemProject[
         languages.Cpp, languages.LanguageWithHeaderFilesSettings, languages.Python
     ]
 ):
@@ -103,7 +103,7 @@ def make_compiledb_factory(
         cache_strategy: cache.Strategy,
     ) -> Compiledb:
         if not source.binding_source:
-            raise otf_compiler.CompilerError(
+            raise compiler.CompilerError(
                 "CMake build system project requires separate bindings code file."
             )
         name = source.program_source.entry_point.name
